requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1       
annotated-types==0.6.0 
anyio==4.2.0
arxiv==2.1.0
async-timeout==4.0.3   
attrs==23.2.0
autogenstudio==0.0.42a0
backoff==2.2.1
beautifulsoup4==4.12.3 
blinker==1.7.0
cachetools==5.3.2      
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
crewai==0.5.5
dataclasses-json==0.6.3
Deprecated==1.2.14
diskcache==5.6.3
distro==1.9.0
docker==7.0.0
exceptiongroup==1.2.0
fastapi==0.109.0
feedparser==6.0.10
FLAML==2.1.1
Flask==2.3.3
frozenlist==1.4.1
google-api-core==2.16.2
google-api-python-client==2.116.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.26.0
idna==3.6
importlib-metadata==6.11.0
itsdangerous==2.1.2
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.0
langchain-community==0.0.16
langchain-core==0.1.17
langchain-openai==0.0.2.post1
langsmith==0.0.85
MarkupSafe==2.1.4
marshmallow==3.20.2
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.3
oauthlib==3.2.2
openai==1.10.0
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-http==1.22.0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
packaging==23.2
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyautogen==0.2.9
pydantic==2.6.0
pydantic_core==2.16.1
pyee==11.1.0
pyparsing==3.1.1
python-dotenv==1.0.1
pywin32==306; sys_platform == 'win32'
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
sgmllib3k==1.0.0
slack_sdk==3.26.2
slackeventsapi==3.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.25
starlette==0.35.1
tenacity==8.2.3
termcolor==2.4.0
tiktoken==0.5.2
tqdm==4.66.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.2.0
uvicorn==0.27.0.post1
waitress==2.1.2
Werkzeug==3.0.1
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0